+++ plugins/UsersManager/Menu.php
@@ -9,7 +9,6 @@
 namespace Piwik\Plugins\UsersManager;
 
 use Piwik\Menu\MenuAdmin;
-use Piwik\Menu\MenuUser;
 use Piwik\Piwik;
 
 class Menu extends \Piwik\Plugin\Menu
@@ -21,12 +20,9 @@ class Menu extends \Piwik\Plugin\Menu
         }
 
         if (Piwik::hasUserSuperUserAccess() && API::getInstance()->getSitesAccessFromUser('anonymous')) {
-            $menu->addSettingsItem('UsersManager_AnonymousUser', $this->urlForAction('anonymousSettings'), $order = 20);
+            $menu->addManageItem('UsersManager_AnonymousUser', $this->urlForAction('anonymousSettings'), $order = 20);
         }
-    }
 
-    public function configureUserMenu(MenuUser $menu)
-    {
         if (!Piwik::isUserIsAnonymous()) {
             $menu->addItem('UsersManager_MenuPersonal', 'General_Settings', $this->urlForAction('userSettings'), 0);
         }